requirement.txt
fastapi
uvicorn
pydantic
python-dotenv
pandas
langchain
langchain-groq
langchain-chroma
langchain-huggingface
langchain-community